Added Notification Receiver
- Added Report Receiver Screen
- Added Report Receiver Service
- Updated Material Module

+++ erp-ranger-admin-portal/src/app/modules/admin/pages/notification-list/notification-list.component.ts
@@ -12,6 +12,7 @@ import { ReportReceiversService } from '../../services/report-receivers.service'
 export class NotificationListComponent implements OnInit {
 
   displayedColumns: string[] = ['importance', 'type', 'sms', 'email', 'web', 'api', 'edit'];
+  displayedColumnsReceivers: string[] = ['park', 'user', 'type', 'edit'];
   reportTypes: ReportType[] = [];
   reportReceivers = [];
   dataSource = new MatTableDataSource(this.reportTypes);
@@ -42,15 +43,15 @@ export class NotificationListComponent implements OnInit {
     });
   }
 
-  deleteReportType(id) {
-    let type;
+  deleteReportType(id: string) {
+    let type: any;
     for (let i = 0; i < this.reportTypes.length; i++) {
       if (this.reportTypes[i].id === id) {
         type = this.reportTypes[i];
         break;
       }
     }
-    if (confirm('Do you want to delete the "' + type.type + '" report type?')) {
+    if (type && confirm('Do you want to delete the "' + type.type + '" report type?')) {
       this.reportTypeService.deleteReportType(id)
         .then(() => {
           alert('Successfully deleted the "' + type.type + '" report type!');
@@ -62,26 +63,45 @@ export class NotificationListComponent implements OnInit {
   }
 
   getReportReceivers() {
-    this.reportReceiverService.getReportTypes().subscribe((results) => {
-      console.log(results);
-      this.reportTypes = [];
+    const park = 'iwGnWNuDC3m1hRzNNBT5';
+    this.reportReceiverService.getReportReceivers(park).subscribe((results) => {
+      this.reportReceivers = [];
       for (let i = 0; i < results.length; i++) {
-        console.log(results[i].payload.doc.data()['user_name']);
-        console.log(results[i].payload.doc.data()['park_name']);
-        const id = results[i].payload.doc.id;
-        this.reportTypeService.readReportType(results[i].payload.doc.id).subscribe((results) => {
-          console.log(id);
-          console.log(results);
+        console.log(results[i].payload.doc.id);
+        this.reportTypeService.readReportType(results[i].payload.doc.data()['type']).subscribe((result) => {
+          this.reportReceivers[i] = {
+            id: String(results[i].payload.doc.id),
+            parkid: String(results[i].payload.doc.data()['park']),
+            userid: String(results[i].payload.doc.data()['user']),
+            typeid: String(results[i].payload.doc.data()['type']),
+            park: String(results[i].payload.doc.data()['park_name']),
+            user: String(results[i].payload.doc.data()['user_name']),
+            type: String(result.payload.data()['type'])
+          };
+          this.dataSourceReceivers = new MatTableDataSource(this.reportReceivers);
         });
-        /*this.reportTypes[i] = {
-          id: String(result[i].payload.doc.id),
-          importance: String(result[i].payload.doc.data()['importance']),
-          methods: result[i].payload.doc.data()['methods'],
-          type: String(result[i].payload.doc.data()['type'])
-        };*/
       }
       this.dataSourceReceivers = new MatTableDataSource(this.reportReceivers);
     });
+  }
+
+  deleteReportReceiver(id: string) {
+    let type: any;
+    for (let i = 0; i < this.reportReceivers.length; i++) {
+      if (this.reportReceivers[i].id === id) {
+        type = this.reportReceivers[i];
+        break;
+      }
+    }
+    if (type && confirm('Do you want to delete the "' + type.type + '" report notification for ' + type.user + '!')) {
+      this.reportReceiverService.deleteReportReceiver(id)
+        .then(() => {
+          alert('Successfully deleted the "' + type.type + '" report notification for ' + type.user + '!');
+        })
+        .catch(() => {
+          alert('Failed to delete the "' + type.type + '" report notification for ' + type.user + '!');
+        });
+    }
   }
 
 }
